Make LunchPadView
